AI_Chat_bot_Gen_AI_retake/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langchain>=0.0.300
langchain-openai>=0.0.1
langchain-community>=0.0.1
PyPDF2>=3.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
PyGithub>=1.59.0
pycryptodome>=3.19.0
